feat: criação do método search na bronze e criação da classe que irão ficar os serviços globais.
